done validation

+++ src/modules/user/user.services.ts
@@ -9,28 +9,40 @@ const createNewUserIntoDB = async (userData: User) => {
 
 //get all user
 const getAllUserFromAllDB = async () => {
-  const result = await userModel.find();
+  const result = await userModel.aggregate([]).project({
+    username: 1,
+    fullName: 1,
+    age: 1,
+    email: 1,
+    address: 1,
+  });
   return result;
 };
 
 // get single user
-const getSingleUserFromAllDB = async (id: string) => {
-  const result = await userModel.findOne({ id });
+const getSingleUserFromAllDB = async (id: number) => {
+  const result = await userModel.findOne({ userId: id });
   return result;
 };
 
 //update user data
 
-const updateSingleUserIntoDB = async (id: string, updatedUserData: User) => {
-  const result = await userModel.findOneAndUpdate({ id }, updatedUserData, {
-    new: true,
-  });
+const updateSingleUserIntoDB = async (id: number, updatedUserData: User) => {
+  const result = await userModel
+    .findOneAndUpdate(
+      { userId: id },
+      { $set: updatedUserData },
+      {
+        new: true,
+      },
+    )
+    .select({ password: 0 });
   return result;
 };
 
 // Delete user data
-const deleteSingleUserFromDB = async (id: string) => {
-  const result = await userModel.findOneAndDelete({ id });
+const deleteSingleUserFromDB = async (id: number) => {
+  const result = await userModel.findOneAndDelete({ userId: id });
   return result;
 };
